Use the same call for storage upload service
